backend/requirements.txt
# Django Core
Django==5.0
djangorestframework==3.14.0
adrf==0.1.6
django-cors-headers==4.3.0

# Database (MySQL instead of PostgreSQL)
mysqlclient==2.2.0
dj-database-url==2.1.0

# Redis & Caching (Compatible with Celery)
redis==4.6.0
django-redis==5.4.0

# Task Queue
celery==5.3.4

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
cryptography==42.0.0

# HTTP Requests (Compatible with python-telegram-bot)
requests==2.31.0
httpx==0.27.0

# Platform SDKs
Telethon==1.36.0
tweepy==4.14.0
twikit==2.3.3
linkedin-api==2.2.0
instagrapi==2.1.2
fbchat==1.9.7
# playwright==1.40.0  # Disabled - too heavy for free tier, using cookie-based auth instead
discord.py==2.3.2

# Validation & Sanitization
bleach==6.1.0

# Server
gunicorn==21.2.0
uvicorn[standard]==0.27.0

# Development
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.92.1
